Trying to improve interactions
Time complexity improve for the best and average case scenario
Some features still need to be worked on like bombs and traps

The logic of interaction changed from every element that wants to
interact having to go through the entire list of elements and calling
for each one the function interact, to a new Logic.
This new logic creates a tmp map<position,list<element>> so that when a
element wants to interact he only needs to check the list of elements in
the position he wants to interact, providing a lot less operations if
the game is running normally.
In exchange for the best case O(n) time complexity we sacrifice O(n)
extra memory, but since this is a relative small program the memory
should hurt a lot